chore: Move new Cypress tests to separate file
